FrontendService/pyproject.toml
[project]
name = "frontendservice"
version = "0.1.0"
description = "Desktop frontend for Finanzuebersicht"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "customtkinter>=5.2,<6",
  "matplotlib>=3.8,<4",
  "numpy>=1.26,<3",
  "python-dotenv>=1.0,<2",
  "requests>=2.31,<3",
  "tkcalendar>=1.6,<2",
]

[project.scripts]
frontendservice = "src.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
